requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
black==21.12b0
bleach==4.1.0
branca==0.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.0.1
django-crispy-forms==1.14.0
django-google-maps==0.12.4
django-ninja==0.16.2
django-ninja-extra==0.14.2
django-ninja-jwt==5.1.6
djongo==1.3.6
dnspython==2.2.0
email-validator==1.1.3
entrypoints==0.3
executing==0.8.2
folium==0.12.1.post1
geographiclib==1.52
geopy==2.2.0
idna==3.3
injector==0.19.0
ipykernel==6.7.0
ipython==8.0.1
ipython-genutils==0.2.0
ipywebrtc==0.6.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
ninja-schema==0.12.8
notebook==6.4.8
numpy==1.22.1
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
pure-eval==0.2.2
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
PyJWT==2.3.0
pymongo==3.7.2
PyMySQL==1.0.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pywin32==303
pywinpty==2.0.1
pyzmq==22.3.0
requests==2.27.1
Send2Trash==1.8.0
six==1.16.0
sqlparse==0.2.4
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
